requirements.txt
asgiref==3.5.0
dj-static==0.0.6
Django==3.2.12
django-stdimage==5.3.0
Pillow==9.0.1
protobuf==3.19.4
pytz==2022.1
sqlparse==0.4.2
static3==0.7.0
